Main Mejorado

+++ src/es/minetsii/MiningCrates/chests/Chest.java
@@ -2,11 +2,11 @@ package es.minetsii.MiningCrates.chests;
 
 public class Chest {
 	private String name;
-	private int probability;
+	private Double probability;
 	private Boolean isConsole;
 	private String command;
 
-	public Chest(String name, int probability, Boolean isConsole, String command) {
+	public Chest(String name, Double probability, Boolean isConsole, String command) {
 		// Nombre;Probabilidad;esConsola;comando
 		// Cofre1;15;true;eco give %p% 5000
 		// Cofre2;5;false;fly
@@ -20,7 +20,7 @@ public class Chest {
 		return name;
 	}
 
-	public int getProbability() {
+	public Double getProbability() {
 		return probability;
 	}
